Added DigitalCoin; Status Output supports more coins
More and more coins are being released and the screen witdh of 80 chars tends to get too small. So in the status output, less chars are used to display all coins. Up to 10 coins should run fine now (puh). If more coins are mined simultaniously, the status output needs to be reworked to support more coins. Todo for the future...